Use a default command line with fallback for sensible commands.
Sensible commands such as ls, cp or chmod may differ a lot between the
BSD and GNU version. This can lead (and already did) to a broken system
when critical installation scripts fail when using non standard options.
We use here a fallback system to the original command when long and
exotic options are used with complex commands.