Adds the fpipe command.
The fpipe command binds the standard input and output to respectively
the write and read parts of a file. This is useful when dealing with
block devices, especially tty. A double buffer is used to ensure that
read and write operations will not block as long as there is space left
in the buffer.