Fixes a bug in ls with very large dates.
When a date for a file was very large the long option of the ls command
caused a segmentation fault. The program didn't take care of the
localtime function giving an EOVERFLOW error hence returning a NULL
pointer. This patch adds a function to parse very large dates with such
overflow error.